+++ packages/article-image/inset-caption.web.js
@@ -1,8 +1,9 @@
+import React from "react";
+import { View } from "react-native";
 import Caption from "@times-components/caption";
 import withResponsiveStyles from "@times-components/responsive-styles";
-import React from "react";
 
-const InsetCaptionStyle = withResponsiveStyles("View", {
+const InsetCaptionStyle = withResponsiveStyles(View, {
   base: () => "padding-left: 10px;",
   mediumUp: () => "padding-left: 0px;"
 });