+++ packages/pull-quote/styles/index.web.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+import sharedStyles from "./shared";
+
+const webStyles = StyleSheet.create({
+  ...sharedStyles,
+  container: {
+    ...sharedStyles.container,
+    marginBottom: 20
+  }
+});
+
+export default webStyles;