+++ packages/pull-quote/styles/shared.js
@@ -1,0 +1,31 @@
+const sharedStyles = {
+  container: {
+    marginTop: 0,
+    marginBottom: 30
+  },
+  quotes: {
+    fontFamily: "TimesModern-Regular",
+    fontSize: 75,
+    marginTop: 0,
+    marginBottom: -40
+  },
+  content: {
+    fontFamily: "TimesModern-Regular",
+    fontSize: 25,
+    lineHeight: 30,
+    color: "#333333"
+  },
+  caption: {
+    fontFamily: "GillSansMTStd-Medium",
+    fontSize: 13,
+    marginTop: 10,
+    marginBottom: 0
+  },
+  link: {
+    color: "#696969",
+    fontFamily: "TimesModern-Bold",
+    textDecorationLine: "none"
+  }
+};
+
+export default sharedStyles;