+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.13.1"></a>
+## [0.13.1](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.13.0...@times-components/article-summary@0.13.1) (2018-01-03)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
+<a name="0.13.0"></a>
+# [0.13.0](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.12.8...@times-components/article-summary@0.13.0) (2018-01-03)
+
+
+### Features
+
+* REPLAT-602 Add Pull Quotes ([#469](https://github.com/newsuk/times-components/issues/469)) ([b2ac5a0](https://github.com/newsuk/times-components/commit/b2ac5a0))
+
+
+
+
 <a name="0.12.8"></a>
 ## [0.12.8](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.12.7...@times-components/article-summary@0.12.8) (2018-01-02)