+++ packages/markup/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.15.1"></a>
+## [0.15.1](https://github.com/newsuk/times-components/compare/@times-components/markup@0.15.0...@times-components/markup@0.15.1) (2018-01-03)
+
+
+
+
+**Note:** Version bump only for package @times-components/markup
+
+<a name="0.15.0"></a>
+# [0.15.0](https://github.com/newsuk/times-components/compare/@times-components/markup@0.14.7...@times-components/markup@0.15.0) (2018-01-03)
+
+
+### Features
+
+* REPLAT-602 Add Pull Quotes ([#469](https://github.com/newsuk/times-components/issues/469)) ([b2ac5a0](https://github.com/newsuk/times-components/commit/b2ac5a0))
+
+
+
+
 <a name="0.14.7"></a>
 ## [0.14.7](https://github.com/newsuk/times-components/compare/@times-components/markup@0.14.6...@times-components/markup@0.14.7) (2018-01-02)